add AnomalibMetric base class
