+++ src/anomalib/models/components/base/anomaly_module.py
@@ -20,11 +20,11 @@ from torchvision.transforms.v2 import Compose, Normalize, Resize, Transform
 
 from anomalib import LearningType
 from anomalib.data import Batch, InferenceBatch
+from anomalib.metrics import F1Score, AUROC
 from anomalib.metrics.threshold import Threshold
 from anomalib.post_processing import OneClassPostProcessor, PostProcessor
 from anomalib.metrics.evaluator import Evaluator
 from anomalib.metrics.collection import MetricWrapper
-from torchmetrics import AUROC, F1Score
 
 from .export_mixin import ExportMixin
 
@@ -218,10 +218,10 @@ class AnomalyModule(ExportMixin, pl.LightningModule, ABC):
 
         Override in subclass for model-specific evaluator behaviour.
         """
-        image_auroc = MetricWrapper(["pred_score", "gt_label"], [AUROC(task="binary")], prefix="image_")
-        image_f1score = MetricWrapper(["pred_label", "gt_label"], [F1Score(task="binary")], prefix="image_")
-        pixel_auroc = MetricWrapper(["anomaly_map", "gt_mask"], [AUROC(task="binary")], prefix="pixel_")
-        pixel_f1score = MetricWrapper(["pred_mask", "gt_mask"], [F1Score(task="binary")], prefix="pixel_")
+        image_auroc = AUROC(fields=["pred_score", "gt_label"], prefix="image_")
+        image_f1score = F1Score(fields=["pred_label", "gt_label"], prefix="image_")
+        pixel_auroc = AUROC(fields=["anomaly_map", "gt_mask"], prefix="pixel_")
+        pixel_f1score = F1Score(fields=["pred_mask", "gt_mask"], prefix="pixel_")
         test_metrics = [image_auroc, image_f1score, pixel_auroc, pixel_f1score]
         return Evaluator(test_metrics=test_metrics)
